controller/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-restx==1.3.0
Flask_JWT_Extended==4.6.0
requests==2.32.2
pyyaml==6.0.2 
bcrypt==4.2.0
psycopg2-binary==2.9.10